refactor(client): split the 3,200-line chat_manager.rs into focused modules
client/src/app/chat_manager.rs had grown into a god file: state, sessions,
contacts, invites, transfers, messaging, and a 600-line event handler in one
3,226-line unit. It is now client/src/app/chat_manager/ with one file per
concern:

  mod.rs       struct + constructor + accessors + toasts + data deletion (~380)
  connect.rs   hosting, connecting (direct/relay/contact), TOFU confirm (~460)
  contacts.rs  contact CRUD, auto-reconnect, group-chat creation       (~160)
  events.rs    session-event pump: poll + handle_session_event         (~620)
  files.rs     transfers: validate, send, receive, wire confirmation   (~250)
  invites.rs   invite links v1/v2/v3 + QR                              (~250)
  text.rs      text send/chunking/reassembly, typing indicators        (~390)
  tests.rs     the unit-test module                                    (~880)

Pure code motion: no public API or behavior change. Cross-module private
helpers are pub(super), scoped to the chat_manager module tree. Docs
(CLAUDE.md, DEVELOPER_GUIDE, 03_architecture) updated to the new layout.
304 workspace tests pass; clippy clean.
